+++ CHANGELOG.md
@@ -2,7 +2,8 @@
   - Added support to Volume actions `:create_snapshot` and `:delete_snapshot`
   - Added support to SAN manager actions
   - Added support to Uplink set actions
-  - Added support to Logical Interconnect
+  - Added support to Logical interconnect
+  - Added support to Server profile actions
 
 ### 0.2.0
   - Upgraded Ruby SDK version to ~> 2.1